Парсинг cli-комманд вынесен в отдельный файл
