+++ src/cli/index.ts
@@ -2,3 +2,5 @@ export * from './cli-application.js';
 export * from './command-parser.js';
 export * from './commands/help.command.js';
 export * from './commands/version.command.js';
+export * from './commands/import.command.js';
+export * from './commands/generate.command.js';
